requirements.txt
cohere
openai
streamlit
python-dotenv
selenium